move proto dir
