+++ impl/authorization/authorization.go
@@ -22,7 +22,7 @@ import (
 	"github.com/google/keytransparency/core/authorization"
 
 	authzpb "github.com/google/keytransparency/core/api/type/type_go_proto"
-	pb "github.com/google/keytransparency/impl/authorization/authz_proto"
+	pb "github.com/google/keytransparency/impl/authorization/authz_go_proto"
 )
 
 type authz struct {
